Add new repositories and update data model
Added several new repositories associated with the game garden package and updated various data models' id field from Int to Long datatype. Additionally, repository interfaces have been updated to extend JpaSpecificationExecutor and PagingAndSortingRepository, enhancing the range of usable operations. Changes to the data model schema also reflected in the database changelog.

+++ src/main/kotlin/de/maurer/gamegarden/model/MemberToGroupKey.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
 @Embeddable
 class MemberToGroupKey (
     @Column(name = "member_id" , nullable = false)
-    val memberId: Int,
+    val memberId: Long,
     @Column(name = "group_id" , nullable = false)
-    val groupId: Int
+    val groupId: Long
 ): Serializable
